feat: modularized page components
