+++ template/web/app/buy-me-coffee/_components/Guide.tsx
