+++ template/web/app/buy-me-coffee/_components/StepOutOfGas.tsx
@@ -1,13 +1,13 @@
 import { useCallback } from 'react';
 import Button from '@/components/Button/Button';
-import { TransactionSteps } from '../ContractDemo';
+import { TransactionSteps } from './ContractDemo';
 
 type OutOfGasStepProps = {
   buyCoffeeAmountRaw: number;
   setTransactionStep: React.Dispatch<React.SetStateAction<TransactionSteps | null>>;
 };
 
-export default function OutOfGasStep({
+export default function StepOutOfGas({
   buyCoffeeAmountRaw,
   setTransactionStep,
 }: OutOfGasStepProps) {
@@ -20,14 +20,11 @@ export default function OutOfGasStep({
       <h2 className="mb-5 w-full text-center text-2xl font-semibold text-white">
         You&apos;re out of gas
       </h2>
-
       <div className="text-center text-6xl">⛽</div>
-
       <div className="my-4 text-center text-sm text-gray-400">
         Please fund your wallet with at least {String(buyCoffeeAmountRaw)} ETH and try sending a
         coffee again.
       </div>
-
       <Button buttonContent="Got it" onClick={handleGotIt} />
     </>
   );
