+++ template/web/app/mint/_components/test/ContractDemo.test.tsx
@@ -3,8 +3,8 @@
  */
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import OnchainProviders from '@/providers/OnchainProviders';
-import ContractDemo from '../../../app/mint/_components/ContractDemo';
+import OnchainProviders from '@/OnchainProviders';
+import ContractDemo from '../ContractDemo';
 
 describe('ContractDemo', () => {
   it("should render fallback if wallet isn't connected", () => {
